Data grids to separate components
