+++ src/backendUrl.ts
@@ -1,4 +1,5 @@
 // const BACKEND_URL = "https://libraryproject.ddns.net";
 const BACKEND_URL = "http://javaohjelmointi.net:15001";
+// const BACKEND_URL = "http://localhost:3000";
 
 export default BACKEND_URL;
