+++ src/components/pages/adminPage/Admin.tsx
@@ -15,6 +15,8 @@ import {
     adminPageTabs,
     adminPageTab
 } from "../../../sxStyles";
+import UsersGrid from "./UsersGrid";
+import BooksGrid from "./BooksGrid";
 
 /*interface IProps {
   books: Array<Book> | undefined;
@@ -30,45 +32,8 @@ interface TabPanelProps {
 }
 
 const Admin: FC = (): JSX.Element => {
-    const [usersData, setUsersData] = useState<User[]>([]);
-    const [booksData, setBooksData] = useState<Book[]>([]);
     const [currentTab, setCurrentTab] = useState<number>(0);
     const navigate = useNavigate();
-
-    const COLUMNS_USERS: GridColDef[] = [
-        { field: "id", headerName: "ID", flex: 2 },
-        { field: "username", headerName: "Username", flex: 2 },
-        { field: "administrator", headerName: "Administrator", flex: 3 },
-        {
-            field: "delete",
-            headerName: "Delete user",
-            flex: 3,
-            renderCell: (params) => (
-                <Button
-                    onClick={() => {
-                        deleteUser(params.row.id);
-                        loadUsersData();
-                    }}
-                >
-                    Delete user
-                </Button>
-            )
-        }
-    ];
-
-    const COLUMNS_BOOKS: GridColDef[] = [
-        { field: "id", headerName: "ID", flex: 1 },
-        { field: "title", headerName: "Title", flex: 5 },
-        { field: "author", headerName: "Author", flex: 4 },
-        { field: "topic", headerName: "Topic", flex: 4 },
-        { field: "location", headerName: "Location", flex: 3 },
-        { field: "isbn", headerName: "ISBN", flex: 3 }
-    ];
-
-    const deleteUser = async (id: number) => {
-        await fetchDeleteUser(id);
-        await loadUsersData();
-    };
 
     const TabPanel = (props: any) => {
         const { children, currentTab, index, ...other } = props;
@@ -79,66 +44,38 @@ const Admin: FC = (): JSX.Element => {
         );
     };
 
-    const loadUsersData = async () => {
-        const usersTmp = await fetchAllUsers();
-        setUsersData(usersTmp);
-    };
-
-    const loadBooksData = async () => {
-        const booksTmp = await fetchAllBooks();
-        setBooksData(booksTmp);
-    };
-
     const handleTabChange = (event: React.SyntheticEvent, newValue: number) => {
         setCurrentTab(newValue);
     };
-
-    useEffect(() => {
-        loadUsersData();
-        loadBooksData();
-    }, []);
-
-    if (usersData.length > 0) {
-        return (
-            <Box sx={{ height: "100vh", width: "100%" }}>
-                <Tabs
-                    value={currentTab}
-                    onChange={handleTabChange}
-                    sx={adminPageTabs}
-                >
-                    <Tab label="Users" sx={adminPageTab} />
-                    <Tab label="Books" sx={adminPageTab} />
-                    <Tab label="Loans" sx={adminPageTab} />
-                </Tabs>
-                <Fab
-                    aria-label="back"
-                    sx={userPageBackButton}
-                    onClick={() => {
-                        navigate("/list-books");
-                    }}
-                >
-                    <ArrowBackIcon />
-                </Fab>
-                <TabPanel index={0} currentTab={currentTab}>
-                    <DataGrid
-                        columns={COLUMNS_USERS}
-                        rows={usersData}
-                        sx={{ width: "100%", height: 1000 }}
-                    ></DataGrid>
-                </TabPanel>
-                <TabPanel index={1} currentTab={currentTab}>
-                    <DataGrid
-                        columns={COLUMNS_BOOKS}
-                        rows={booksData}
-                        sx={{ width: "100%", height: 1000 }}
-                    ></DataGrid>
-                </TabPanel>
-                <TabPanel index={2} currentTab={currentTab}></TabPanel>
-            </Box>
-        );
-    } else {
-        return <></>;
-    }
+    return (
+        <Box sx={{ height: "100vh", width: "100%" }}>
+            <Tabs
+                value={currentTab}
+                onChange={handleTabChange}
+                sx={adminPageTabs}
+            >
+                <Tab label="Users" sx={adminPageTab} />
+                <Tab label="Books" sx={adminPageTab} />
+                <Tab label="Loans" sx={adminPageTab} />
+            </Tabs>
+            <Fab
+                aria-label="back"
+                sx={userPageBackButton}
+                onClick={() => {
+                    navigate("/list-books");
+                }}
+            >
+                <ArrowBackIcon />
+            </Fab>
+            <TabPanel index={0} currentTab={currentTab}>
+                <UsersGrid />
+            </TabPanel>
+            <TabPanel index={1} currentTab={currentTab}>
+                <BooksGrid />
+            </TabPanel>
+            <TabPanel index={2} currentTab={currentTab}></TabPanel>
+        </Box>
+    );
 };
 
 export default Admin;
